Refactored Avatar Modal
